pyproject.toml
[project]
name = "langgraph-sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.7",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.10",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.1.0",
    "plotly>=5.17.0",
    "pillow>=10.0.0",
]
